requirements.txt
pandas>=1.3.0
openpyxl>=3.0.9
numpy>=1.20.0
python-dateutil>=2.8.2
pytz>=2021.1
SQLAlchemy>=1.4.0
matplotlib>=3.4.0
seaborn>=0.11.0